move tests to single place
